Create test base
